+++ src/vscode-util.ts
@@ -126,10 +126,12 @@ type PaperbackWriterConfiguration = {
     header: {
       items: HeaderFooterItems[]
       fontSize: HeaderFooterFOntSize
+      margin: string
     }
     footer: {
       items: HeaderFooterItems[]
       fontSize: HeaderFooterFOntSize
+      margin: string
     }
     printBackground: boolean
     paperOrientation: PdfOrientation
@@ -204,11 +206,13 @@ export const getPaperbackWriterConfiguration = (scope?: vscode.ConfigurationScop
       displayHeaderFooter: wsc['PDF']['displayHeaderFooter'],
       header: {
         items: wsc['PDF']['header']['items'],
-        fontSize: wsc['PDF']['header']['fontSize']
+        fontSize: wsc['PDF']['header']['fontSize'],
+        margin: wsc['PDF']['header']['margin']
       },
       footer: {
         items: wsc['PDF']['footer']['items'],
-        fontSize: wsc['PDF']['footer']['fontSize']
+        fontSize: wsc['PDF']['footer']['fontSize'],
+        margin: wsc['PDF']['footer']['margin']
       },
       printBackground: wsc['PDF']['printBackground'],
       paperOrientation: wsc['PDF']['paperOrientation'],